tests: fold temporal regression tests into thematic test classes
Moves the 6 tests from test_temporal_regressions.py into existing
test_temporal.py classes:
- trusted_roots_refresh + worker_interceptor_accepts_provider → TestTenuoPluginConfig
- latency_ring_is_bounded → TestTenuoMetrics
- three _resolve_warrant_id_* tests → new TestWorkflowInboundWarrantId
  (first coverage of signal/update denial paths)
